requirements.txt
amqp==2.6.1
asgiref==3.4.1
astor==0.8.1
backcall==0.2.0
billiard==3.6.4.0
celery==4.4.1
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
decorator==5.1.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.10
django-author==1.0.4
django-debug-toolbar==3.2.4
django-import-export==2.7.1
django-import-export-celery==1.1.6
djangorestframework==3.13.1
et-xmlfile==1.1.0
flake8==3.9.2
flake8-django==1.1.2
flake8-django-migrations==0.1.1
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.10.0
ipython==7.30.1
jedi==0.18.1
kombu==4.6.11
MarkupPy==1.14
matplotlib-inline==0.1.3
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.9
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.0
prompt-toolkit==3.0.24
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pycodestyle==2.7.0
pyflakes==2.3.1
Pygments==2.10.0
python-decouple==3.6
pytz==2021.3
PyYAML==6.0
redis==3.4.1
requests==2.27.1
setuptools-git==1.2
sqlparse==0.4.2
tablib==3.1.0
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.8
vine==1.3.0
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
zipp==3.7.0
